agent_requirements.txt
# OSU RAG Query Agent — Dependencies
# Requires Python 3.10+

# --- Google ADK (with A2A protocol support) ---
google-adk[a2a]

# --- Google GenAI (embeddings) ---
google-genai>=1.5.0

# --- Vector Database ---
pinecone>=5.4.2

# --- Environment ---
python-dotenv>=1.0.1
